+++ quietdrop-tauri/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "quietdrop-frontend"
+name = "quietdrop-tauri"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,7 +7,7 @@ edition = "2021"
 yew = { version = "0.20", features = ["csr"] }
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
-js-sys = "0.3"
-web-sys = { version = "0.3", features = ["HtmlInputElement"] }
+web-sys = { version = "0.3", features = ["HtmlInputElement", "console"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+gloo-utils = "0.1"
