+++ quietdrop-tauri/index.html
@@ -3,8 +3,7 @@
 <head>
     <meta charset="utf-8" />
     <title>QuietDrop</title>
-    <link data-trunk rel="scss" href="styles/main.scss" />
-    <!-- <link data-trunk rel="copy-dir" href="public" /> -->
+    <link data-trunk rel="css" href="styles.css" />
 </head>
 <body>
     <div id="app"></div>
