+++ quietdrop-tauri/src-tauri/.gitignore
