+++ quietdrop-tauri/src-tauri/capabilities/default.json
