+++ quietdrop-tauri/src-tauri/icons/128x128.png
