+++ quietdrop-tauri/src-tauri/icons/128x128@2x.png
