+++ quietdrop-tauri/src-tauri/icons/Square107x107Logo.png
