+++ quietdrop-tauri/src-tauri/icons/Square142x142Logo.png
