+++ quietdrop-tauri/src-tauri/icons/Square150x150Logo.png
