+++ quietdrop-tauri/src-tauri/icons/Square284x284Logo.png
