+++ quietdrop-tauri/src-tauri/icons/Square30x30Logo.png
