+++ quietdrop-tauri/src-tauri/icons/Square310x310Logo.png
