+++ quietdrop-tauri/src-tauri/icons/Square44x44Logo.png
