+++ quietdrop-tauri/src-tauri/icons/Square71x71Logo.png
