+++ quietdrop-tauri/src-tauri/icons/StoreLogo.png
