+++ quietdrop-tauri/src-tauri/icons/icon.icns
