+++ quietdrop-tauri/src-tauri/icons/icon.ico
