+++ quietdrop-tauri/src-tauri/icons/icon.png
