+++ quietdrop-tauri/src-tauri/tauri.conf.json
@@ -1,10 +1,11 @@
 {
-  "$schema": "https://raw.githubusercontent.com/tauri-apps/tauri/2.0.0-alpha/core/tauri-config-schema/schema.json",
-  "identifier": "com.quietdrop.app",
+  "$schema": "https://schema.tauri.app/config/2",
   "productName": "QuietDrop",
   "version": "0.1.0",
+  "identifier": "com.tauri.dev",
   "build": {
-    "frontendDist": "../dist"
+    "frontendDist": "../dist",
+    "devUrl": "http://localhost:1420"
   },
   "app": {
     "windows": [
@@ -22,6 +23,13 @@
   },
   "bundle": {
     "active": true,
-    "targets": "all"
+    "targets": "all",
+    "icon": [
+      "icons/32x32.png",
+      "icons/128x128.png",
+      "icons/128x128@2x.png",
+      "icons/icon.icns",
+      "icons/icon.ico"
+    ]
   }
 }
