--- quietdrop-ui/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8" />
-    <title>QuietDrop</title>
-    <link data-trunk rel="css" href="styles.css" />
-</head>
-<body>
-    <div id="app"></div>
-</body>
-</html>
