--- quietdrop-ui/src-tauri/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "app"
-version = "0.1.0"
-description = "A Tauri App"
-authors = ["you"]
-license = ""
-repository = ""
-edition = "2021"
-rust-version = "1.77.2"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-name = "app_lib"
-crate-type = ["staticlib", "cdylib", "rlib"]
-
-[build-dependencies]
-tauri-build = { version = "2.1.1", features = [] }
-
-[dependencies]
-quietdrop-core = { path = "../../quietdrop-core" }
-serde_json = "1.0"
-serde = { version = "1.0", features = ["derive"] }
-log = "0.4"
-tauri = { version = "2.4.1", features = [] }
-tauri-plugin-log = { version = "2.0.0-rc", features = ["colored"] }
-tauri-plugin-shell = "2.0.0-rc"
-tokio = { version = "1.x", features = ["full"] }
-chrono = { version = "0.4.40", features = ["serde"] }
-sodiumoxide = "0.2.7"
