--- quietdrop-ui/src-tauri/tauri.conf.json
@@ -1,35 +1,0 @@
-{
-  "$schema": "https://schema.tauri.app/config/2",
-  "productName": "QuietDrop",
-  "version": "0.1.0",
-  "identifier": "com.tauri.dev",
-  "build": {
-    "frontendDist": "../dist",
-    "devUrl": "http://localhost:1420"
-  },
-  "app": {
-    "windows": [
-      {
-        "title": "QuietDrop",
-        "width": 800,
-        "height": 600,
-        "resizable": true,
-        "fullscreen": false
-      }
-    ],
-    "security": {
-      "csp": null
-    }
-  },
-  "bundle": {
-    "active": true,
-    "targets": "all",
-    "icon": [
-      "icons/32x32.png",
-      "icons/128x128.png",
-      "icons/128x128@2x.png",
-      "icons/icon.icns",
-      "icons/icon.ico"
-    ]
-  }
-}
